Replace use of deprecated numpy aliases for python int, float, bool.

+++ astropy/io/fits/tests/test_compression_failures.py
@@ -10,7 +10,7 @@ from . import FitsTestCase
 
 
 MAX_INT = np.iinfo(np.intc).max
-MAX_LONG = np.iinfo(np.long).max
+MAX_LONG = np.iinfo(int).max
 MAX_LONGLONG = np.iinfo(np.longlong).max
